requirements.txt
# AI SDKs
google-genai>=1.0.0
openai>=1.0.0

# Face validation
insightface>=0.7.3
albumentations>=1.3.0

# ML
torch>=2.0.0
numpy<2

# Image
Pillow>=10.0.0

# HTTP
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Backend (FastAPI)
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic>=2.8.0
PyJWT>=2.9.0
ffmpeg-python>=0.2.0
